requirements.txt
pymupdf
opencv-python
pyzbar
pytesseract
pandas
matplotlib
numpy
google-genai
python-dotenv
Pillow
